backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-nmap==0.7.1
scapy==2.5.0
httpx==0.27.0
